refactor: move agent controller factory to controller layer
